Bugfix: keep a copy of the original authorized_keys
The ansible assemble command builds a new authorized_keys
file for root.
Before this fix, it used to overwrite the original authorized_keys,
so ansible cannot access the machine anymore, unless using the same
certificate.

Tried to implement with the copy module, but remote_src is only
supported in 2.0, so implemented using the copy command